refactor(components): migrate CloseButton to TypeScript
